refactor: 型情報をtypesに移行した
