+++ src/apis/blogs/index.ts
@@ -1,16 +1,7 @@
 import { MicroCMSQueries } from 'microcms-js-sdk';
 
-import { MicroCMSList, MicroCMSObject } from 'apis/MicroCMSTypes';
-import { Tag } from 'apis/tags';
-
-export type Blog = MicroCMSObject & {
-  title: string;
-  author: ('okazaki' | 'kobayashi' | 'satou' | 'kinoshita' | 'ryou' | 'suda')[];
-  recommend: boolean;
-  category: ('engineer' | 'design' | 'column')[];
-  tags: Tag[];
-  body: string;
-};
+import { Blog } from 'types/blog';
+import { MicroCMSList } from 'types/microCMS';
 
 export type Methods = {
   get: {
