+++ src/apis/news/index.ts
@@ -1,14 +1,7 @@
 import { MicroCMSQueries } from 'microcms-js-sdk';
 
-import { MicroCMSList, MicroCMSObject } from 'apis/MicroCMSTypes';
-
-export type News = MicroCMSObject & {
-  title: string;
-  important: boolean;
-  category: string[];
-  notificationDate: Date;
-  body: string;
-};
+import { MicroCMSList } from 'types/microCMS';
+import { News } from 'types/news';
 
 export type Methods = {
   get: {
