+++ src/types/tag.ts
@@ -1,0 +1,6 @@
+import { MicroCMSObject } from 'types/microCMS';
+
+export type Tag = MicroCMSObject & {
+  nameEn: string;
+  nameJa: string;
+};
